+++ standalone/build.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
 
-pyinstaller ./rakaia.spec --noconfirm
+pyinstaller ./rakaia.spec --noconfirm --clean
 chmod 777 dist/*
 chmod +x dist/*